Fixed Bug
A bug was found involving the default sensor. No if a sensor is default then it is forced to be visible even when an attempt to change it is made. Also, you may not remove the default sensor without it anothe sensor being made default. This prevents fron end bugs.