Updated Security Configuration
The Bean for the clientId and clientSecret was buggy. It seemed to be related to a Spring problem. To avoid it and to better secure the two values as well as make it easy to change them in the future, a reader was implemented for a security.properties file. (Not included on github for obvious reasons).