Set up domain configuration
To fill in the html for the navigation to point to the right domain it was added to a configuration file and implemented. In the future it is best to get away from this method of injecting the domain if possible. This gets glitchy with loading, though there are updates that fix these problems it is best to reduce the amount of "moving parts".