Updated navigation bar loading
Needed to make changes to all page builders to allow the super class to handle the Model objects. this allows for easier acces to the attributes that contribute to the navigation bar.

This push is also in preperation for an addition of more endpoints and pages that will be added in the coming weeks.