Updated Configuration Reader to be a singleton
This reader was implemented in multiple places. The file was being reread for "no reason". Everything in certain pages were read. Once this is deployed the configuration file does not change very often, therefore we do not need to worry about it being re-read. Thats why a singleton was implemented.