Major Authorization Push
Authorization push that is the barebones for how authorization will take place in the back end. Authorization is not yet implemented completely and this push is expected to break the web app until the database is set up for users. (Break it by no longer responding correctly to signed in users on the navigation bar)

More updates will come soon